requirements.txt
Pillow==10.2.0
pytesseract==0.3.10
streamlit==1.28.1
